auth.deleteMyCredentials
