+++ internal/wrappers/cpp/auth.cpp
@@ -32,4 +32,11 @@ namespace kuzzleio {
     delete(r);
     return ret;
   }
+
+  void Auth::deleteMyCredentials(const std::string& strategy, query_options *options) Kuz_Throw_KuzzleException {
+    void_result *r = kuzzle_delete_my_credentials(_auth, const_cast<char*>(strategy.c_str()), options);
+    if (r->error != NULL)
+        throwExceptionFromStatus(r);
+    delete(r);
+  }
 }
